parser: swap all log.Errorf calls with error values, take 2
Second take, as the first merge was undone by a bot.

As I was working on my patch for #70, I was wondering if we really
needed a logger interface with both Infof and Errorf. After all, a
library should report its errors as values, not via a logger.

Turns out it's relatively straightforward to do so. All functions that
used to call log.Errorf are themselves called from top-level APIs which
are able to return an error, so we just need to thread them.

One special case is getMetaCommentStringValues, which used to report
each error and continue rather than stop. We currently do not have an
idiomatic way to report multiple errors at once, so return the first
error for now. In the future, we could use https://go.dev/issue/53435.

Updates #70.