parser: replace glog Infof calls with an interface
Now that glog.Errorf calls are returned as error values,
we can replace glog.Verbose.Infof calls with an interface.

All calls go through a logger in the config.
By default the logger is nil, which disables logging,
so we have nil checks when logging.

Note that the library uses a nil check on the logger to tell whether or
not it can avoid constructing the log messages.
We don't need a method like "IsInfoLevel" or "IsEnabled",
since the interface only knows how to log at the info level.

With that in mind, keep the behavior in cmd/txtpbfmt unchanged;
it now passes log.V(2) in the config if its level is enabled.

Fixes #70.